% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_auxilliary_data.R
\name{compute_auxilliary_data}
\alias{compute_auxilliary_data}
\title{Compute group-wise data on top of the raw data}
\usage{
compute_auxilliary_data(data_in)
}
\arguments{
\item{data_in, }{containing columns "group_id", "offered", "treated", "outcome",
where data$group_id are positive integers, other entries in data all (0,1)}
}
\value{
data_out, sorted by group_id,
  with additional columns: "group_size", "frac_treated_exclude_i", "frac_compilers_exclude_i"
  The value of "frac_compilers_exclude_i" will be NaN for zero-saturation groups
}
\description{
For each individual i belonging to group g,
  compute the total size of the group g,
  compute the fraction (excluding i) of treated individual in group g,
  compute the fraction (excluding i) of compilers in group g,
Assume data does not have NA.
Assume that in each non-zero saturation group there are at least 2 offered.
}
