% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_instruments.R
\name{compute_instruments}
\alias{compute_instruments}
\title{Compute instruments needed for identification}
\usage{
compute_instruments(data_in, moments_of_s)
}
\arguments{
\item{data_in}{A data frame, containing columns "group_id", "offered",
"treated", "outcome",group_size", "frac_treated_exclude_i",
"frac_compilers_exclude_i".}

\item{moments_of_s}{A matrix, containing parameters related to the
randomised saturation experimental design, computed by the function
compute_moments_of_s}
}
\value{
A data frame with containing the original data and additional columns
  "cal_ZW" containing 4 by 4 matrices
  "cal_Z1_for_ii" containing 2 by 2 matrices
  "cal_Z1_for_iii" containing 2 by 2 matrices
}
\description{
Compute instruments needed for identification of treatment effects
with spill-over effect and
randomised saturation experiment and partial_compliance.
Assume each group have at least two people offered treatment
and the data does not contain NA.
}
\examples{
test_data
moments_of_s
compute_instruments(test_data, moments_of_s)
}
