% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_instruments_robust.R
\name{compute_instruments_robust}
\alias{compute_instruments_robust}
\title{Compute instruments needed for identification, tolerate zero saturation designs}
\usage{
compute_instruments_robust(data_in, moments_of_s)
}
\arguments{
\item{data_in}{A data frame, containing columns "group_id", "offered",
"treated", "outcome",group_size", "frac_treated_exclude_i",
"frac_compilers_exclude_i".}

\item{moments_of_s}{A matrix, containing parameters related to the
randomised saturation experimental design, computed by the function
compute_moments_of_s}
}
\value{
A data frame with containing the original data and additional columns
  "cal_ZW" containing 5/4 by 1 matrices, 5 if contains data-contains zero saturations
  "cal_Z1_for_ii" containing 2 by 1 matrices
  "cal_Z1_for_iii" containing 2 by 1 matrices
}
\description{
Compute instruments needed for identification of treatment effects
with spill-over effect and
randomised saturation experiment and partial_compliance.
Assume each non-zero saturation group have at least two people offered treatment
and the data does not contain NA.
}
