% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_moment_of_s.R
\name{compute_moments_of_s}
\alias{compute_moments_of_s}
\title{Compute expectations of [s^(i-1)][(1-s)^(j-1)]}
\usage{
compute_moments_of_s(val_s, prob_s)
}
\arguments{
\item{val_s}{The set of values of s}

\item{prob_s}{The probability of s equals to corresponding entries in val_s}
}
\value{
moments_of_s, a matrix of doubles whose (i,j) entry stores E[s^(i-1)][(1-s)^(j-1)]
}
\description{
Compute expectations of [s^(i-1)][(1-s)^(j-1)]
for (i,j) = (1,2), (2,1), (2,2), (2,3), (3,1), (3,2), (4,1)
}
