% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_Qs.R
\name{compute_Q_0}
\alias{compute_Q_0}
\title{Compute Q_0 according to the formula in Appedix C, step 3}
\usage{
compute_Q_0(moments_of_s, frac_compilers_exclude_i, group_size)
}
\description{
Compute Q_0 according to the formula in Appedix C, step 3
}
