% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_Qs.R
\name{compute_Q_1_robust}
\alias{compute_Q_1_robust}
\title{Compute Q_1 for possibly zero saturation, return NA if saturation = 0.}
\usage{
compute_Q_1_robust(
  moments_of_s,
  frac_compilers_exclude_i,
  group_size,
  total_offered
)
}
\description{
Compute Q_1 for possibly zero saturation, return NA if saturation = 0.
}
