% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_Qs.R
\name{compute_Q_robust}
\alias{compute_Q_robust}
\title{Compute Q for possibly zero saturation, return NA if saturation = 0.}
\usage{
compute_Q_robust(Q_0, Q_1, total_offered)
}
\description{
Compute Q for possibly zero saturation, return NA if saturation = 0.
}
