% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_treatment_effects.R
\name{compute_treatment_effects}
\alias{compute_treatment_effects}
\title{Compute parameters that identify the treatment effects}
\usage{
compute_treatment_effects(data_in, val_s, prob_s)
}
\arguments{
\item{data_in}{A data frame, containing columns "group_id", "offered",
"treated", "outcome"}

\item{val_s}{Vector of saturation levels}

\item{prob_s}{Vector of saturation assignment probabilities}
}
\value{
A data frame with estimates and standard errors of various coeffcients
In particular
  \eqn{\gamma^e} is the population average average negative spill-over rate
  \eqn{\gamma^n} is the non-compilers average negative spill-over rate
  \eqn{\beta^c + \delta^c d_bar} is the average compiler direct effect,
  where d_bar is the avergae number of treated group neighbours.
  Various other treatment effects on the treated/untreated can be deduced
  from these coefficients, see Theorem 3 of the paper and its proof in the appendix.
}
\description{
Compute parameters that identify the treatment effects
}
\examples{

}
