% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_Zs.R
\name{compute_Z1_for_ii}
\alias{compute_Z1_for_ii}
\title{Compute the instrument Z1 for (ii) in Theorem 2 as a function of Q1, treated,
and fraction_treated in group (excluding the current data point)}
\usage{
compute_Z1_for_ii(Q_1, treated, frac_treated_exclude_i)
}
\description{
Compute the instrument Z1 for (ii) in Theorem 2 as a function of Q1, treated,
and fraction_treated in group (excluding the current data point)
}
