% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_Zs.R
\name{compute_Z1_for_ii_robust}
\alias{compute_Z1_for_ii_robust}
\title{Compute Z1 for (ii) for designs with zero-saturation groups,
if saturation is zero return NA.}
\usage{
compute_Z1_for_ii_robust(Q_1, treated, frac_treated_exclude_i, total_offered)
}
\description{
Compute Z1 for (ii) for designs with zero-saturation groups,
if saturation is zero return NA.
}
