% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_Zs.R
\name{compute_Z1_for_iii}
\alias{compute_Z1_for_iii}
\title{Compute the instrument Z1 for (iii) in Theorem 2 as a function of Q2, offered,
treated, and fraction_treated in group (excluding the current data point)}
\usage{
compute_Z1_for_iii(Q_1, offered, treated, frac_treated_exclude_i)
}
\description{
Compute the instrument Z1 for (iii) in Theorem 2 as a function of Q2, offered,
treated, and fraction_treated in group (excluding the current data point)
}
