% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_Zs.R
\name{compute_ZW}
\alias{compute_ZW}
\title{Compute the instrument ZW for (i) in Theorem 2 as a function of Q, offered,
and fraction_treated in group (excluding the current data point)}
\usage{
compute_ZW(Q, offered, frac_treated_exclude_i)
}
\description{
Compute the instrument ZW for (i) in Theorem 2 as a function of Q, offered,
and fraction_treated in group (excluding the current data point)
}
