% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_Zs.R
\name{compute_ZW_robust}
\alias{compute_ZW_robust}
\title{Compute ZW for (i) in Theorem 2, for designs with zero-saturation groups,
using the formula in Appendix D.}
\usage{
compute_ZW_robust(Q, offered, frac_treated_exclude_i, total_offered)
}
\description{
Compute ZW for (i) in Theorem 2, for designs with zero-saturation groups,
using the formula in Appendix D.
}
