% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_saturations.R
\name{transform_saturations}
\alias{transform_saturations}
\title{Compute conditioning (on saturation being non-zero) probabilities of being assigned a saturation level.}
\usage{
transform_saturations(val_s, prob_s)
}
\arguments{
\item{val_s, }{values of original saturation levels.}

\item{prob_s, }{probability of being assigned each original saturation levels.}
}
\value{
list(val_s_excl_zero, prob_s_excl_zero), transformed saturation levels
and probabilities.
}
\description{
Given the design of a randomised saturation experiment with a zero saturation
group, compute saturation values and probabilities conditioning on the
saturation being non-zero.
}
